+++ src/cem-parser.ts
@@ -5,16 +5,18 @@ import {
   getMemberDescription,
   removeQuotes,
 } from "@wc-toolkit/cem-utilities";
-import type { ArgTypes, ControlOptions } from "./storybook-types";
-import type { Options } from "./types";
+import type { ArgTypes } from "@storybook/web-components";
+import type { ControlOptions } from "./storybook-types.js";
+import type { StorybookHelpersOptions } from "./types.js";
 import type { Component } from "@wc-toolkit/cem-utilities";
+import { CssCustomProperty } from "custom-elements-manifest";
 
 type ArgSet = {
   resets?: ArgTypes;
   args: ArgTypes;
 };
 
-let options: Options = {};
+let options: StorybookHelpersOptions = {};
 
 setTimeout(() => {
   options = (globalThis as any)?.__WC_STORYBOOK_HELPERS_CONFIG__ || {};
@@ -22,7 +24,7 @@ setTimeout(() => {
 
 export function getAttributesAndProperties(
   component?: Component,
-  enabled = true
+  enabled = true,
 ): {
   resets?: ArgTypes;
   propArgs: ArgTypes;
@@ -38,7 +40,7 @@ export function getAttributesAndProperties(
     }
 
     const attribute = component.attributes?.find(
-      (x) => member.name === x.fieldName
+      (x) => member.name === x.fieldName,
     );
     const propName = member.name;
     const args = attribute ? attrArgs : propArgs;
@@ -63,7 +65,9 @@ export function getAttributesAndProperties(
       ? (member as any)[`${options.typeRef}`]?.text || member?.type?.text
       : member?.type?.text;
     const propType = cleanUpType(type);
-    const defaultValue = member.readonly ? undefined : removeQuotes(member.default || "");
+    const defaultValue = member.readonly
+      ? undefined
+      : removeQuotes(member.default || "");
     const control = getControl(propType, attribute !== undefined);
 
     args[name] = {
@@ -71,20 +75,17 @@ export function getAttributesAndProperties(
       description: getDescription(
         member.description,
         propName,
-        member.deprecated as string
+        member.deprecated as string,
       ),
       defaultValue: defaultValue
-          ? defaultValue === "''"
-            ? ""
-            : control === "object"
-              ? JSON.parse(formatToValidJson(defaultValue))
-              : defaultValue
-          : undefined,
-      control: enabled && !member.readonly
-        ? {
-            type: control,
-          }
-        : false,
+        ? defaultValue === "''"
+          ? ""
+          : control === "object"
+            ? JSON.parse(formatToValidJson(defaultValue))
+            : defaultValue
+        : undefined,
+      control:
+        enabled && !member.readonly && control ? { type: control } : false,
       table: {
         category: attribute ? "attributes" : "properties",
         defaultValue: {
@@ -107,7 +108,7 @@ export function getAttributesAndProperties(
 
 export function getReactProperties(
   component?: Component,
-  enabled = true
+  enabled = true,
 ): ArgSet {
   const resets: ArgTypes = {};
   const args: ArgTypes = {};
@@ -143,11 +144,10 @@ export function getReactProperties(
       name: member.name,
       description: member.description,
       defaultValue: getDefaultValue(controlType, member.default),
-      control: enabled && !member.readonly
-        ? {
-            type: controlType,
-          }
-        : false,
+      control:
+        enabled && !member.readonly && controlType
+          ? { type: controlType }
+          : false,
       table: {
         category: "properties",
         defaultValue: {
@@ -191,7 +191,7 @@ export function getReactEvents(component?: Component): ArgSet {
 
 export function getCssProperties(
   component?: Component,
-  enabled = true
+  enabled = true,
 ): ArgSet {
   const resets: ArgTypes = {};
   const args: ArgTypes = {};
@@ -212,9 +212,7 @@ export function getCssProperties(
       defaultValue: property.default,
       control: enabled
         ? {
-            type: property.name?.toLowerCase()?.includes("color")
-              ? "color"
-              : "text",
+            type: getCssPropControl(property),
           }
         : false,
       table: {
@@ -224,6 +222,26 @@ export function getCssProperties(
   });
 
   return { resets, args };
+}
+
+function getCssPropControl(
+  property: CssCustomProperty,
+): ControlOptions | undefined {
+  const type = (property as any).type?.text?.toLowerCase();
+  const name = property.name?.toLowerCase();
+  if (
+    name?.includes("color") ||
+    name?.includes("colour") ||
+    type === "<color>"
+  ) {
+    return "color";
+  }
+
+  if (type === "<integer>" || type === "<number>") {
+    return "number";
+  }
+
+  return "text";
 }
 
 export function getCssParts(component?: Component, enabled = true): ArgSet {
@@ -242,7 +260,7 @@ export function getCssParts(component?: Component, enabled = true): ArgSet {
       name: part.name,
       description: getDescription(
         part.description,
-        enabled ? `${part.name}-part` : ""
+        enabled ? `${part.name}-part` : "",
       ),
       control: enabled ? "text" : false,
       table: {
@@ -270,7 +288,7 @@ export function getCssStates(component?: Component, enabled = true): ArgSet {
       name: state.name,
       description: getDescription(
         state.description,
-        enabled ? `${state.name}-state` : ""
+        enabled ? `${state.name}-state` : "",
       ),
       control: enabled ? "text" : false,
       table: {
@@ -299,7 +317,7 @@ export function getSlots(component?: Component, enabled = true): ArgSet {
       name: slotName,
       description: getDescription(
         slot.description,
-        enabled ? `${slotName}-slot` : ""
+        enabled ? `${slotName}-slot` : "",
       ),
       control: enabled ? "text" : false,
       table: {
@@ -440,7 +458,7 @@ function cleanUpType(type?: string): string {
 function getDescription(
   description?: string,
   argRef?: string,
-  deprecated?: string
+  deprecated?: string,
 ) {
   let desc = getMemberDescription(description, deprecated);