+++ src/storybook-helpers.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { TemplateResult } from "lit";
-import { getStyleTemplate, getTemplate } from "./html-templates.js";
+import { getStyleTemplate, getTemplate, logEvent } from "./html-templates.js";
 import {
   getCssParts,
   getCssProperties,
@@ -13,13 +13,16 @@ import {
   getMethods,
 } from "./cem-parser.js";
 import { Component, getComponentByTagName } from "@wc-toolkit/cem-utilities";
-import type { ArgTypes } from "./storybook-types";
-import type { Categories, Options, StoryHelpers, StoryOptions } from "./types";
+import type { ArgTypes } from "@storybook/web-components";
+import type {
+  Categories,
+  StorybookHelpersOptions,
+  StoryHelpers,
+  StoryOptions,
+} from "./types.js";
 import type { Package } from "custom-elements-manifest";
 
-let userOptions: Options =
-  (globalThis as any)?.__WC_STORYBOOK_HELPERS_CONFIG__ || {};
-const defaultOptions: Options = {
+const defaultOptions: StorybookHelpersOptions = {
   typeRef: "parsedType",
   categoryOrder: [
     "attributes",
@@ -33,11 +36,18 @@ const defaultOptions: Options = {
   ],
 };
 
+let userOptions: StorybookHelpersOptions = {
+  ...defaultOptions,
+  ...((globalThis as any)?.__WC_STORYBOOK_HELPERS_CONFIG__ || {}),
+};
+
 /**
  * sets the global config for the Storybook helpers
  * @param options
  */
-export function setStorybookHelpersConfig(options: Options) {
+export function setStorybookHelpersConfig(
+  options: StorybookHelpersOptions = {},
+) {
   options = { ...defaultOptions, ...options };
   (globalThis as any).__WC_STORYBOOK_HELPERS_CONFIG__ = options;
   userOptions = options;
@@ -50,9 +60,12 @@ export function setStorybookHelpersConfig(options: Options) {
  */
 export function getStorybookHelpers<T>(
   tagName: string,
-  options?: StoryOptions
+  options?: StoryOptions,
 ): StoryHelpers<T> {
-  userOptions = (globalThis as any)?.__WC_STORYBOOK_HELPERS_CONFIG__ || {};
+  userOptions = {
+    ...defaultOptions,
+    ...((globalThis as any)?.__WC_STORYBOOK_HELPERS_CONFIG__ || {}),
+  };
   const cem = getManifest();
   const component = getComponent(cem, tagName);
   const eventNames = component?.events?.map((event) => event.name) || [];
@@ -64,6 +77,7 @@ export function getStorybookHelpers<T>(
     reactArgs: getReactArgs(component),
     reactArgTypes: getReactProps(component),
     events: eventNames,
+    logEvent,
     styleTemplate: (args?: any) =>
       getStyleTemplate(component, args, options?.excludeCategories || []),
     template: (args?: any, slot?: TemplateResult) =>
@@ -73,7 +87,7 @@ export function getStorybookHelpers<T>(
         slot,
         argTypes,
         options?.excludeCategories || [],
-        options?.setComponentVariable
+        options?.setComponentVariable,
       ),
   };
 
@@ -85,7 +99,7 @@ function getManifest(): Package {
   const cem: Package = (window as any).__STORYBOOK_CUSTOM_ELEMENTS_MANIFEST__;
   if (!cem) {
     throw new Error(
-      `Custom Elements Manifest not found. Be sure to follow the pre-install steps in this guide:\nhttps://www.npmjs.com/package/wc-storybook-helpers#before-you-install`
+      `Custom Elements Manifest not found. Be sure to follow the pre-install steps in this guide:\nhttps://www.npmjs.com/package/wc-storybook-helpers#before-you-install`,
     );
   }
   return cem;
@@ -96,7 +110,7 @@ function getComponent(cem: Package, tagName: string): Component | undefined {
 
   if (!component) {
     throw new Error(
-      `A component with the tag name "${tagName}" was not found in the Custom Elements Manifest. If it's missing in the CEM, it's often the result of a missing "@tag" or "@tagName" tag in the component's JSDoc.\nAdditional information can be found here:\nhttps://custom-elements-manifest.open-wc.org/analyzer/getting-started/#supported-jsdoc`
+      `A component with the tag name "${tagName}" was not found in the Custom Elements Manifest. If it's missing in the CEM, it's often the result of a missing "@tag" or "@tagName" tag in the component's JSDoc.\nAdditional information can be found here:\nhttps://custom-elements-manifest.open-wc.org/analyzer/getting-started/#supported-jsdoc`,
     );
   }
 
@@ -110,7 +124,7 @@ function getComponent(cem: Package, tagName: string): Component | undefined {
  */
 function getArgTypes(
   component?: Component,
-  excludeCategories?: Array<Categories>
+  excludeCategories?: Array<Categories>,
 ): ArgTypes {
   const cssProps = getCssProperties(component);
   const cssParts = getCssParts(component);
@@ -141,7 +155,7 @@ function getArgTypes(
     attrsAndProps.resets,
     events.resets,
     cssStates.resets,
-    methods.resets
+    methods.resets,
   );
 
   getCategoriesOrder().forEach((category) => {
@@ -158,8 +172,8 @@ function getArgTypes(
  * @param argTypes argTypes object for component
  * @returns an object containing the `args` for the component
  */
-function getArgs<T>(argTypes: ArgTypes): Partial<T> {
-  const args: Partial<T> = {};
+function getArgs<T>(argTypes: ArgTypes): Partial<T> & { [key: string]: any } {
+  const args: Partial<T> & { [key: string]: any } = {};
   for (const [key, value] of Object.entries(argTypes)) {
     if (value?.control) {
       args[key as keyof T] = getDefaultValue(value.defaultValue) || "";
@@ -191,7 +205,7 @@ function getDefaultValue(value?: string | number | boolean | object) {
  */
 function getReactProps(
   component?: Component,
-  excludeCategories?: Array<Categories>
+  excludeCategories?: Array<Categories>,
 ): ArgTypes {
   const cssProps = getCssProperties(component);
   const cssParts = getCssParts(component);
@@ -200,6 +214,10 @@ function getReactProps(
   const events = getReactEvents(component);
   const cssStates = getCssStates(component);
   const methods = getMethods(component);
+  const options: StorybookHelpersOptions = {
+    ...defaultOptions,
+    ...((globalThis as any)?.__WC_STORYBOOK_HELPERS_CONFIG__ || {}),
+  };
 
   const args: Record<Exclude<Categories, "attributes">, ArgTypes> = {
     cssParts: cssParts.args,
@@ -225,7 +243,7 @@ function getReactProps(
     (category) => {
       if (excludeCategories?.includes(category)) return;
       argTypes = { ...argTypes, ...(args[category] || {}) };
-    }
+    },
   );
 
   return sortByCategories(argTypes, {...args, attributes: {}});